pyproject.toml
[project]
name = "paralysis"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "avulto>=0.1.5",
    "click>=8.1.7",
    "ipython>=8.32.0",
    "largestinteriorrectangle>=0.2.1",
    "llvmlite==0.43.0",
    "loguru>=0.7.2",
    "mysqlclient>=2.2.4",
    "numba==0.60.0",
    "numpy==2.0",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=10.4.0",
    "rasterio>=1.4.0",
    "requests-cache>=1.2.1",
    "requests-ratelimiter>=0.7.0",
    "seaborn>=0.13.2",
    "shapely>=2.0.6",
    "sqlalchemy>=2.0.35",
    "typed-settings>=24.5.0",
]

[project.scripts]
space_ruin_map = "paralysis.tools.space_ruin_map:main"
lavaland_ruin_map = "paralysis.tools.lavaland_ruin_map:main"
wiki_areamap = "paralysis.tools.wiki_areamap:main"
testmerges = "paralysis.blackbox.testmerges:main"
sync_blackbox = "paralysis.tools.sync_blackbox:main"
create_tables = "paralysis.tools.create_tables:main"
map_stitch = "paralysis.tools.map_stitch:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
